REF: Move c_shared_lib() into extern module
